pyproject.toml
[tool.poetry]
name = "respo"
version = "0.1.0"
description = "Resource Policy in Python made easy"
license = "MIT"
readme = "README.md"
homepage = "https://github.com/rafsaf/respo"
documentation = "https://respo.rafsaf.pl"
repository = "https://github.com/rafsaf/respo"
authors = ["rafsaf <rafal.safin@rafsaf.pl>"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Security",
    "Topic :: Internet",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Internet :: WWW/HTTP",
    "Typing :: Typed",
]
[tool.poetry.dependencies]
python = "^3.6.1"
pydantic = "^1.8.2"
PyYAML = "^5.4.1"
typer = "^0.4.0"
colorama = "^0.4.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
coverage = {extras = ["toml"], version = "^5.5"}
isort = "^5.9.3"
black = "^20.7b0"

[tool.coverage.run]
source = [
    "respo"
]

[tool.isort]
profile = "black"
known_third_party = ["respo"]
skip_glob = [
    "respo/__init__.py",
    ]

[tool.poetry.scripts]
respo = "respo.typer:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

